# config/swagger.yaml
swagger: '2.0'
info:
  title: BasePlatform
  description: A NodeJS Platform as a Service
  version: 1.0.0
schemes:
  - http
  - https
basePath: /api/v1
produces:
  - application/json
paths:
...
  '/applications/{appId}/deployments':
    get:
      summary: Deployment List
      description: |
        Get list of deployments
      parameters:
        - name: appId
          in: path
          description: Application ID
          required: true
          type: string   
      tags:
        - Deployments
      responses:
        '200':
          description: An array of deployment objects
          schema:
            type: array
            items:
              $ref: '#/definitions/Deployment'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create Deployment
      description: |
        Create Deployment
      parameters:
        - name: appId
          in: path
          description: Application ID
          required: true
          type: string   
        - name: gitRepository
          in: formData
          description: GIT Repository
          required: false
          type: string
        - name: zipLocation
          in: formData
          description: ZIP Location
          required: false
          type: string
        - name: zip
          in: formData
          description: ZIP File
          required: false
          type: file          
      tags:
        - Deployments
      responses:
        '200':
          description: A success object
          schema:
            $ref: '#/definitions/Success'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/applications/{appId}/deployments/{deployId}':
    get:
      summary: Deployment Details
      description: |
        Get details of a deployment
      parameters:
        - name: appId
          in: path
          description: Application ID
          required: true
          type: string
        - name: deployId
          in: path
          description: Deployment ID
          required: true
          type: string
      tags:
        - Deployments
      responses:
        '200':
          description: A deployment object
          schema:
            type: array
            items:
              $ref: '#/definitions/Deployment'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Update Deployment
      description: |
        Update Deployment
      parameters:
        - name: appId
          in: path
          description: Application ID
          required: true
          type: string     
        - name: deployId
          in: path
          description: Deploy ID
          required: true
          type: string       
        - name: gitRepository
          in: formData
          description: GIT Repository
          required: false
          type: string
        - name: zipLocation
          in: formData
          description: ZIP Location
          required: false
          type: string
        - name: zip
          in: formData
          description: ZIP File
          required: false
          type: file   
      tags:
        - Deployments
      responses:
        '200':
          description: A success object
          schema:
            $ref: '#/definitions/Success'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete Deployment
      description: |
        Delete Deployment
      parameters:
        - name: appId
          in: path
          description: Application ID
          required: true
          type: string 
        - name: deployId
          in: path
          description: Deployment ID
          required: true
          type: string  
      tags:
        - Deployments
      responses:
        '200':
          description: A success object
          schema:
            $ref: '#/definitions/Success'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/machines':
    get:
      summary: Machine List
      description: |
        Get list of machines
      tags:
        - Machines
      responses:
        '200':
          description: An array of machine objects
          schema:
            type: array
            items:
              $ref: '#/definitions/Machine'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create Machine
      description: |
        Create Application
      parameters:
        - name: name
          in: formData
          description: Machine Name
          required: false
          type: string
        - name: description
          in: formData
          description: Machine Description
          required: false
          type: string
        - name: machineUrl
          in: formData
          description: Machine URL
          required: false
          type: string
      tags:
        - Machines
      responses:
        '200':
          description: A success object
          schema:
            $ref: '#/definitions/Success'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Deployment:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier
      gitRepository:
        type: string
        description: GIT Repository
      zipLocation:
        type: string
        description: ZIP Location
      zipUploaded:
        type: string
        description: ZIP Uploaded
  Machine:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier
      name:
        type: string
        description: Machine Name
      description:
        type: string
        description: Machine Description
      machineUrl:
        type: string
        description: Machine URL
...
  Success:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier of object
      type:
        type: string
        description: 'Type of object (user, event, etc)'
      result:
        type: string
        description: Success
      action:
        type: string
        description: 'Create, Update, Delete or [Other]'
      shortMessage:
        type: string
        description: Short message describing success
      longMessage:
        type: string
        description: Long message describing success
      supplementary:
        type: object
        description: Contains supplementary data for the request
  Error:
    type: object
    properties:
      result:
        type: string
        description: Error
      code:
        type: integer
        format: int32
        description: Error code
      shortMessage:
        type: string
        description: Short error message
      longMessage:
        type: string
        description: Long error message
      correlationId:
        type: string
        description: Log ID used for error tracking and troubleshooting